backend/requirements.txt
fastapi==0.113.0
uvicorn[standard]==0.23.2
langchain==1.2.15
langchain-openai==1.1.12
openai>=2.26.0,<3.0.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
langchain-openai==1.1.12
tiktoken>=0.4.0,<1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6